+++ crates/prelude/src/crypto.rs
@@ -16,6 +16,7 @@
 
 pub mod ccm;
 pub mod gcm;
+pub mod hash;
 
 /// Errors returned by cryptographic operations.
 pub use wasefire_applet_api::crypto::Error;