+++ crates/api-macro/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.3.0-git
+
+### Major
+
+- Update `wasefire-applet-api-desc` version
+
 ## 0.2.0
 
 ### Major